# src/main/resources/static/openapi.yaml
openapi: 3.1.0

info:
  title: Online Store REST API
  version: 1.0.0
  description: |
    A teaching API for exploring customers, cities, products, purchases,
    refunds, and store summary information.

    The current REST implementation returns hardcoded stub objects.
    The JPA model and CSV import exercise are included in the same project
    so students can later replace the stub service with repository queries.

servers:
  - url: http://localhost:8080
    description: Local Spring Boot server

tags:
  - name: API
    description: API discovery
  - name: Customers
    description: Customer information and customer activity
  - name: Cities
    description: Store activity grouped by city
  - name: Products
    description: Product information
  - name: Purchases
    description: Purchase information
  - name: Refunds
    description: Refund information
  - name: Summary
    description: Store-level summary information

paths:

  /api:
    get:
      tags: [API]
      summary: List API resources
      operationId: getApiIndex
      responses:
        "200":
          description: Links to the main API resources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiIndex"
              example:
                customers: /api/customers
                cities: /api/cities
                products: /api/products
                purchases: /api/purchases
                refunds: /api/refunds
                summary: /api/summary

  /api/customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Return all customers, or restrict the result to a city.
      operationId: getCustomers
      parameters:
        - $ref: "#/components/parameters/CityQuery"
      responses:
        "200":
          description: Customers matching the request.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
              example:
                - id: 1
                  name: Alice Chen
                  zip: "02108"
                  city: Boston
                - id: 10
                  name: Jamal Ferris
                  zip: "02116"
                  city: Boston

  /api/customers/{customerId}:
    get:
      tags: [Customers]
      summary: Get a customer
      operationId: getCustomer
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: The requested customer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/customers/{customerId}/purchases:
    get:
      tags: [Customers, Purchases]
      summary: List purchases for a customer
      operationId: getCustomerPurchases
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Purchases made by the customer.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"

  /api/customers/{customerId}/refunds:
    get:
      tags: [Customers, Refunds]
      summary: List refunds for a customer
      operationId: getCustomerRefunds
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Refunds issued to the customer.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Refund"

  /api/cities:
    get:
      tags: [Cities]
      summary: List cities
      operationId: getCities
      responses:
        "200":
          description: Cities represented by the customers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
              example:
                - name: Atlanta
                  customerCount: 1
                - name: Boston
                  customerCount: 2

  /api/cities/{city}/customers:
    get:
      tags: [Cities, Customers]
      summary: List customers in a city
      operationId: getCityCustomers
      parameters:
        - $ref: "#/components/parameters/CityPath"
      responses:
        "200":
          description: Customers in the requested city.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"

  /api/cities/{city}/purchases:
    get:
      tags: [Cities, Purchases]
      summary: List purchases in a city
      operationId: getCityPurchases
      parameters:
        - $ref: "#/components/parameters/CityPath"
      responses:
        "200":
          description: Purchases associated with customers in the city.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"

  /api/cities/{city}/refunds:
    get:
      tags: [Cities, Refunds]
      summary: List refunds in a city
      operationId: getCityRefunds
      parameters:
        - $ref: "#/components/parameters/CityPath"
      responses:
        "200":
          description: Refunds associated with customers in the city.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Refund"

  /api/products:
    get:
      tags: [Products]
      summary: List products
      description: Return all current products, or filter by part of a product name.
      operationId: getProducts
      parameters:
        - $ref: "#/components/parameters/ProductNameQuery"
      responses:
        "200":
          description: Products matching the request.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /api/products/{productId}:
    get:
      tags: [Products]
      summary: Get a product
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: The requested product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/purchases:
    get:
      tags: [Purchases]
      summary: List purchases
      description: Return all purchases, or filter by customer ID and/or city.
      operationId: getPurchases
      parameters:
        - $ref: "#/components/parameters/CustomerIdQuery"
        - $ref: "#/components/parameters/CityQuery"
      responses:
        "200":
          description: Purchases matching the request.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"

  /api/purchases/{purchaseId}:
    get:
      tags: [Purchases]
      summary: Get a purchase
      operationId: getPurchase
      parameters:
        - $ref: "#/components/parameters/PurchaseId"
      responses:
        "200":
          description: The requested purchase.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/refunds:
    get:
      tags: [Refunds]
      summary: List refunds
      description: Return all refunds, or filter by customer ID and/or city.
      operationId: getRefunds
      parameters:
        - $ref: "#/components/parameters/CustomerIdQuery"
        - $ref: "#/components/parameters/CityQuery"
      responses:
        "200":
          description: Refunds matching the request.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Refund"

  /api/refunds/{refundId}:
    get:
      tags: [Refunds]
      summary: Get a refund
      operationId: getRefund
      parameters:
        - $ref: "#/components/parameters/RefundId"
      responses:
        "200":
          description: The requested refund.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refund"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/summary:
    get:
      tags: [Summary]
      summary: Get a store summary
      operationId: getStoreSummary
      responses:
        "200":
          description: Store-level counts and totals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSummary"
              example:
                customers: 10
                cities: 9
                purchases: 8
                products: 9
                refunds: 5
                purchaseTotal: 404.39
                refundTotal: 220.45

components:

  parameters:

    CustomerId:
      name: customerId
      in: path
      required: true
      description: Unique customer identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1

    CustomerIdQuery:
      name: customerId
      in: query
      required: false
      description: Return records belonging to this customer.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1

    ProductId:
      name: productId
      in: path
      required: true
      description: Unique product identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 101

    PurchaseId:
      name: purchaseId
      in: path
      required: true
      description: Unique purchase identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1001

    RefundId:
      name: refundId
      in: path
      required: true
      description: Unique refund identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 9001

    CityPath:
      name: city
      in: path
      required: true
      description: City name.
      schema:
        type: string
      example: Boston

    CityQuery:
      name: city
      in: query
      required: false
      description: Restrict results to this city.
      schema:
        type: string
      example: Boston

    ProductNameQuery:
      name: name
      in: query
      required: false
      description: Case-insensitive text contained in the product name.
      schema:
        type: string
      example: mouse

  responses:

    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  schemas:

    ApiIndex:
      type: object
      additionalProperties:
        type: string
      description: Map of resource names to API paths.

    Customer:
      type: object
      required: [id, name, zip, city]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Alice Chen
        zip:
          type: string
          example: "02108"
        city:
          type: string
          example: Boston

    City:
      type: object
      required: [name, customerCount]
      properties:
        name:
          type: string
          example: Boston
        customerCount:
          type: integer
          format: int64
          example: 2

    Product:
      type: object
      required: [id, name, quantity, price]
      properties:
        id:
          type: integer
          format: int64
          example: 101
        name:
          type: string
          example: Wireless Mouse
        quantity:
          type: integer
          minimum: 0
          example: 1
        price:
          type: number
          format: double
          minimum: 0
          example: 24.99

    Purchase:
      type: object
      required: [id, date, customerId, customerName, city, products, total]
      properties:
        id:
          type: integer
          format: int64
          example: 1001
        date:
          type: string
          format: date
          example: "2026-01-05"
        customerId:
          type: integer
          format: int64
          example: 1
        customerName:
          type: string
          example: Alice Chen
        city:
          type: string
          example: Boston
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
          format: double
          example: 34.98

    Refund:
      type: object
      required:
        [id, date, customerId, customerName, city, productName, quantity, price, amount, refundType]
      properties:
        id:
          type: integer
          format: int64
          example: 9001
        date:
          type: string
          format: date
          example: "2026-01-12"
        customerId:
          type: integer
          format: int64
          example: 1
        customerName:
          type: string
          example: Alice Chen
        city:
          type: string
          example: Boston
        productName:
          type: string
          example: USB-C Cable
        quantity:
          type: integer
          minimum: 1
          example: 1
        price:
          type: number
          format: double
          minimum: 0
          example: 9.99
        amount:
          type: number
          format: double
          minimum: 0
          example: 9.99
        refundType:
          type: string
          enum: [PARTIAL, FULL]
          example: PARTIAL

    StoreSummary:
      type: object
      required:
        [customers, cities, purchases, products, refunds, purchaseTotal, refundTotal]
      properties:
        customers:
          type: integer
          format: int64
          example: 10
        cities:
          type: integer
          format: int64
          example: 9
        purchases:
          type: integer
          format: int64
          example: 8
        products:
          type: integer
          format: int64
          example: 9
        refunds:
          type: integer
          format: int64
          example: 5
        purchaseTotal:
          type: number
          format: double
          example: 404.39
        refundTotal:
          type: number
          format: double
          example: 220.45

    Problem:
      type: object
      description: Standard HTTP problem response.
      additionalProperties: true